+++ inst/scripts/assay-functions-Ex.R
@@ -20,7 +20,6 @@ qreduceAssay(re4, query, weightedmean)
 \dontrun{
     ## Extended example: non-silent mutations, summarized by genic
     ## region
-
     suppressPackageStartupMessages({
         library(TxDb.Hsapiens.UCSC.hg19.knownGene)
         library(org.Hs.eg.db)
@@ -30,11 +29,36 @@ qreduceAssay(re4, query, weightedmean)
         library(TCGAutils)
     })
 
-    ## TCGA Multi-assay experiment to RaggedExperiment
-
-
+    ## TCGA MultiAssayExperiment with RaggedExperiment data
     mae <- curatedTCGAData("ACC", c("RNASeq2GeneNorm", "CNASNP", "Mutation"),
         dry.run = FALSE)
 
-    simplifyTCGA(mae)
+    ## genomic coordinates
+    gn <- genes(TxDb.Hsapiens.UCSC.hg19.knownGene)
+    gn <- keepStandardChromosomes(granges(gn), pruning.mode="coarse")
+    seqlevelsStyle(gn) <- "NCBI"
+    gn <- unstrand(gn)
+
+    ## reduce mutations, marking any genomic range with non-silent
+    ## mutation as FALSE
+    nonsilent <- function(scores, ranges, qranges)
+        any(scores != "Silent")
+    mre <- mae[["ACC_Mutation-20160128"]]
+    genome(mre) <- translateBuild(genome(re))
+    mutations <- qreduceAssay(mre, gn, nonsilent, "Variant_Classification")
+
+    ## reduce copy number
+    re <- mae[["ACC_CNASNP-20160128"]]
+    class(re)
+    ## [1] "RaggedExperiment"
+    genome(re) <- "hg19"
+    cn <- qreduceAssay(re, gn, weightedmean, "Segment_Mean")
+
+    ## ALTERNATIVE
+    ##
+    ## TCGAutils helper function to convert RaggedExperiment objects to
+    ## RangedSummarizedExperiment based on annotated gene ranges
+    mae[[1L]] <- re
+    mae[[2L]] <- mre
+    qreduceTCGA(mae)
 }